feat(infrastructure): add function profiling
The profiler has been modified to take snapshots of the counters and at regular intervals and to
leverage systemtap to figure out what function is currently in execution.

BREAKING CHANGE: the data that will be generated will now include the function name as a new
feature. moreover a single execution of the target program will now create many snapshots. Each
snapshot is a row in the dataset.

Signed-off-by: Mattia Nicolella <mattianicolella@gmail.com>